Add client-side wasm version of Recast
