+++ package.json
@@ -51,7 +51,6 @@
     ]
   },
   "dependencies": {
-    "@donmccurdy/recast": "0.0.4",
     "@fortawesome/fontawesome-free": "^5.2.0",
     "@mozillareality/react-dnd-html5-backend": "^5.1.11",
     "@robertlong/react-ui-tree": "github:robertlong/react-ui-tree#dev",
@@ -108,6 +107,7 @@
     "babel-preset-env": "^1.7.0",
     "babel-preset-react": "^6.24.1",
     "browser-env": "^3.2.5",
+    "copy-webpack-plugin": "^4.5.2",
     "cross-env": "^5.2.0",
     "css-loader": "^0.28.11",
     "dotenv": "^6.0.0",
@@ -117,10 +117,8 @@
     "eslint-plugin-react": "^7.8.2",
     "esm": "^3.0.79",
     "file-loader": "^1.1.11",
-    "glob": "^7.1.3",
     "html-webpack-plugin": "^3.2.0",
     "husky": "^0.14.3",
-    "jszip": "^3.1.5",
     "koa-webpack": "^5.0.2",
     "lint-staged": "^7.2.0",
     "node-sass": "^4.9.1",
